fix(web): TOC 停靠栏内联与弹窗并存 + 拖拽方向 + 状态保持
- 抽离 TocPanel 纯内容组件：TocDrawer 变薄壳（BottomSheet），TocDock 直接复用
  TocPanel，修复宽屏下内联停靠栏与底部弹窗同时出现
- 拖拽方向修正：dock 右缘固定、左缘跟随指针，向左拖增宽向右拖变窄
- 分界线交互/样式对齐 SplitDivider：setPointerCapture、hover 扩展 grab 区、
  accent 高亮、pointercancel 清理
- TOC 开合状态全局保持：切换/关闭文件不再重置（移除 currentFile watch 与
  closeOverlayAndSync 中的 close）
