+++ web/src/components/__tests__/TocDrawer.test.ts
@@ -1,6 +1,6 @@
-import { describe, expect, it, vi, beforeEach } from 'vitest'
+import { describe, expect, it, vi } from 'vitest'
 import { mount } from '@vue/test-utils'
-import { nextTick, defineComponent } from 'vue'
+import { defineComponent } from 'vue'
 import TocDrawer from '@/components/TocDrawer.vue'
 
 // ── Mocks ──
@@ -10,10 +10,6 @@ vi.mock('vue-i18n', () => ({
     t: (key: string) => {
       const map: Record<string, string> = {
         'toc.title': 'TOC',
-        'toc.searchPlaceholder': 'Search...',
-        'toc.loading': 'Loading...',
-        'toc.noMatch': 'No match',
-        'toc.noHeadings': 'No headings',
       }
       return map[key] ?? key
     },
@@ -41,66 +37,16 @@ vi.mock('@/components/common/HeaderMarquee.vue', () => ({
   }),
 }))
 
-vi.mock('@/components/common/SearchInput.vue', () => ({
+vi.mock('@/components/TocPanel.vue', () => ({
   default: defineComponent({
-    props: { modelValue: String, placeholder: String },
-    emits: ['update:modelValue', 'enter', 'down', 'up', 'dblclick'],
-    template: '<input class="search-input-stub" :value="modelValue" @input="$emit(\'update:modelValue\', $event.target.value)" />',
+    name: 'TocPanel',
+    props: { file: Object, pdfOutline: Array },
+    emits: ['jump', 'jumpPage'],
+    template: '<div class="toc-panel-stub" />',
   }),
 }))
 
-vi.mock('@/components/common/LoadingIndicator.vue', () => ({
-  default: defineComponent({
-    props: { label: String, size: String },
-    template: '<div class="loading-indicator-stub">{{ label }}</div>',
-  }),
-}))
-
-vi.mock('@/composables/useCodeSymbols', () => ({
-  fetchCodeSymbols: vi.fn(() => Promise.resolve(null)),
-}))
-
-// Mutable mock for isEditorDirty so we can toggle it per test
-let mockIsEditorDirty = false
-
-vi.mock('@/composables/useFileEditor', () => ({
-  useFileEditor: () => ({
-    isEditorDirty: () => mockIsEditorDirty,
-  }),
-}))
-
-vi.mock('@/composables/useListNav', () => ({
-  useListNav: () => ({
-    activeIndex: { value: -1 },
-    reset: vi.fn(),
-    confirm: vi.fn(),
-    down: vi.fn(),
-    up: vi.fn(),
-  }),
-}))
-
-vi.mock('@/composables/useListKeys', () => ({
-  useListKeys: () => {},
-}))
-
-vi.mock('@/utils/fileType.ts', () => ({
-  getFileType: (name: string) => ({
-    lang: name.endsWith('.md') ? 'markdown' : 'go',
-    isMarkdown: name.endsWith('.md'),
-  }),
-}))
-
-// Mock IntersectionObserver (not available in jsdom)
-class MockIntersectionObserver {
-  observe() {}
-  disconnect() {}
-  unobserve() {}
-}
-beforeEach(() => {
-  vi.stubGlobal('IntersectionObserver', MockIntersectionObserver)
-})
-
-function mountTocDrawer(props: Record<string, any> = {}) {
+function mountDrawer(props: Record<string, any> = {}) {
   return mount(TocDrawer, {
     props: {
       open: true,
@@ -117,248 +63,42 @@ function mountTocDrawer(props: Record<string, any> = {}) {
 }
 
 describe('TocDrawer', () => {
-  it('renders empty state when no file is provided', () => {
-    const wrapper = mountTocDrawer({ file: null })
-    expect(wrapper.find('.toc-empty').text()).toContain('No headings')
+  it('renders the bottom sheet wrapper with a TOC header', () => {
+    const wrapper = mountDrawer({ file: { name: 'a.md', path: '/a.md' } })
+    expect(wrapper.find('.bottom-sheet').exists()).toBe(true)
+    expect(wrapper.find('.bs-header').exists()).toBe(true)
   })
 
-  it('renders empty state when file has no content', () => {
-    const wrapper = mountTocDrawer({ file: { name: 'test.md', content: '' } })
-    expect(wrapper.find('.toc-empty').exists()).toBe(true)
+  it('shows the file path in the header when present', () => {
+    const wrapper = mountDrawer({ file: { name: 'a.md', path: '/a.md' } })
+    expect(wrapper.find('.header-marquee').text()).toContain('/a.md')
   })
 
-  it('renders markdown headings from file content', async () => {
-    const wrapper = mountTocDrawer({
-      file: { name: 'readme.md', content: '# Title\n## Section 1\n### Sub', path: '/readme.md' },
-    })
-    await nextTick()
-    await nextTick()
-    const items = wrapper.findAll('.toc-item')
-    expect(items.length).toBeGreaterThanOrEqual(1)
-  })
-
-  it('shows PDF outline when pdfOutline is provided', async () => {
-    const wrapper = mountTocDrawer({
-      file: { name: 'doc.pdf', path: '/doc.pdf' },
-      pdfOutline: [
-        { id: 'p1', text: 'Page 1', level: 1, line: 1 },
-        { id: 'p2', text: 'Page 2', level: 2, line: 2 },
-      ],
-    })
-    await nextTick()
-    const items = wrapper.findAll('.toc-item')
-    expect(items.length).toBe(2)
-    // Should show page badge
-    expect(wrapper.find('.toc-page-badge').exists()).toBe(true)
-  })
-
-  it('emits jumpPage when clicking PDF outline item', async () => {
-    const wrapper = mountTocDrawer({
-      file: { name: 'doc.pdf', path: '/doc.pdf' },
-      pdfOutline: [{ id: 'p5', text: 'Page 5', level: 1, line: 5 }],
-    })
-    await nextTick()
-    await wrapper.find('.toc-item').trigger('click')
-    expect(wrapper.emitted('jumpPage')).toBeTruthy()
-    expect(wrapper.emitted('jumpPage')![0]).toEqual([5])
-  })
-
-  it('emits close when clicking a PDF outline item', async () => {
-    const wrapper = mountTocDrawer({
-      file: { name: 'doc.pdf', path: '/doc.pdf' },
+  it('renders TocPanel with the file and pdfOutline', () => {
+    const wrapper = mountDrawer({
+      file: { name: 'a.md', path: '/a.md' },
       pdfOutline: [{ id: 'p1', text: 'Page 1', level: 1, line: 1 }],
     })
-    await nextTick()
-    await wrapper.find('.toc-item').trigger('click')
+    const panel = wrapper.findComponent({ name: 'TocPanel' })
+    expect(panel.exists()).toBe(true)
+    expect(panel.props('file')).toMatchObject({ name: 'a.md' })
+    expect(panel.props('pdfOutline')).toEqual([{ id: 'p1', text: 'Page 1', level: 1, line: 1 }])
+  })
+
+  it('forwards jump from TocPanel and emits close (bottom-sheet: tap to dismiss)', () => {
+    const wrapper = mountDrawer({ file: { name: 'a.md', path: '/a.md' } })
+    const panel = wrapper.findComponent({ name: 'TocPanel' })
+    panel.vm.$emit('jump', 12)
+    expect(wrapper.emitted('jump')).toEqual([[12]])
+    // Bottom-sheet TOC closes after jumping to an item.
     expect(wrapper.emitted('close')).toBeTruthy()
   })
 
-  it('emits jumpPage but does NOT emit close when docked', async () => {
-    const wrapper = mountTocDrawer({
-      file: { name: 'doc.pdf', path: '/doc.pdf' },
-      pdfOutline: [{ id: 'p5', text: 'Page 5', level: 1, line: 5 }],
-      docked: true,
-    })
-    await nextTick()
-    await wrapper.find('.toc-item').trigger('click')
-    expect(wrapper.emitted('jumpPage')).toBeTruthy()
-    expect(wrapper.emitted('close')).toBeFalsy()
-  })
-
-  it('emits jump but does NOT emit close when docked for text headings', async () => {
-    const wrapper = mountTocDrawer({
-      file: { name: 'readme.md', content: '# Title\n## Section 1', path: '/readme.md' },
-      docked: true,
-    })
-    await nextTick()
-    await nextTick()
-    await wrapper.find('.toc-item').trigger('click')
-    expect(wrapper.emitted('jump')).toBeTruthy()
-    expect(wrapper.emitted('close')).toBeFalsy()
-  })
-})
-
-describe('TocDrawer — editor dirty path', () => {
-  it('uses client-side extractToc when editor is dirty', async () => {
-    mockIsEditorDirty = true
-
-    const { fetchCodeSymbols } = await import('@/composables/useCodeSymbols')
-    const spy = vi.mocked(fetchCodeSymbols)
-    spy.mockClear()
-
-    const wrapper = mountTocDrawer({
-      file: { name: 'main.go', content: 'func main() {}', path: '/main.go' },
-    })
-    await nextTick()
-    await nextTick()
-    // fetchCodeSymbols should NOT be called because editor is dirty
-    expect(spy).not.toHaveBeenCalled()
-    // Should still have items from regex extraction
-    const items = wrapper.findAll('.toc-item')
-    expect(items.length).toBeGreaterThanOrEqual(0)
-
-    mockIsEditorDirty = false
-  })
-})
-
-describe('TocDrawer — fetchCodeSymbols with results', () => {
-  it('renders code symbols from fetchCodeSymbols', async () => {
-    const { fetchCodeSymbols } = await import('@/composables/useCodeSymbols')
-    vi.mocked(fetchCodeSymbols).mockResolvedValueOnce({
-      lang: 'go',
-      symbols: [
-        { name: 'main', kind: 'function', line: 10, endLine: 20, level: 1 },
-        { name: 'Handler', kind: 'struct', line: 25, endLine: 40, level: 1 },
-      ],
-    })
-
-    const wrapper = mountTocDrawer({
-      file: { name: 'main.go', content: 'package main', path: '/main.go' },
-    })
-    await nextTick()
-    // Wait for async fetchCodeSymbols
-    await new Promise(r => setTimeout(r, 50))
-    await nextTick()
-    const items = wrapper.findAll('.toc-item')
-    expect(items.length).toBe(2)
-  })
-
-  it('falls back to extractToc when fetchCodeSymbols returns null', async () => {
-    const { fetchCodeSymbols } = await import('@/composables/useCodeSymbols')
-    vi.mocked(fetchCodeSymbols).mockResolvedValueOnce(null)
-
-    const wrapper = mountTocDrawer({
-      file: { name: 'main.go', content: 'package main', path: '/main.go' },
-    })
-    await nextTick()
-    await new Promise(r => setTimeout(r, 50))
-    await nextTick()
-    // extractToc for go with no matching patterns should give empty
-    const items = wrapper.findAll('.toc-item')
-    // Go regex fallback may not extract from plain text
-    expect(items.length).toBeGreaterThanOrEqual(0)
-  })
-
-  it('falls back to extractToc when fetchCodeSymbols throws', async () => {
-    const { fetchCodeSymbols } = await import('@/composables/useCodeSymbols')
-    vi.mocked(fetchCodeSymbols).mockRejectedValueOnce(new Error('network'))
-
-    const wrapper = mountTocDrawer({
-      file: { name: 'main.go', content: 'package main', path: '/main.go' },
-    })
-    await nextTick()
-    await new Promise(r => setTimeout(r, 50))
-    await nextTick()
-    const items = wrapper.findAll('.toc-item')
-    expect(items.length).toBeGreaterThanOrEqual(0)
-  })
-
-  it('deduplicates heading IDs from code symbols', async () => {
-    const { fetchCodeSymbols } = await import('@/composables/useCodeSymbols')
-    vi.mocked(fetchCodeSymbols).mockResolvedValueOnce({
-      lang: 'markdown',
-      symbols: [
-        { name: 'Intro', kind: 'heading', line: 1, endLine: 1, level: 2 },
-        { name: 'Intro', kind: 'heading', line: 10, endLine: 10, level: 2 },
-        { name: 'Details', kind: 'heading', line: 20, endLine: 20, level: 2 },
-      ],
-    })
-
-    const wrapper = mountTocDrawer({
-      file: { name: 'doc.md', content: '## Intro\n## Intro\n## Details', path: '/doc.md' },
-    })
-    await nextTick()
-    await new Promise(r => setTimeout(r, 50))
-    await nextTick()
-    const items = wrapper.findAll('.toc-item')
-    expect(items.length).toBe(3)
-    // First Intro gets base ID, second gets -2 suffix
-    expect(items[0].attributes('class')).toContain('toc-item')
-  })
-})
-
-describe('TocDrawer — watch cancellation', () => {
-  it('cancels fetchCodeSymbols result when file changes before resolve', async () => {
-    const { fetchCodeSymbols } = await import('@/composables/useCodeSymbols')
-    let resolveFirst: (v: any) => void
-    const firstCall = new Promise(r => { resolveFirst = r })
-    vi.mocked(fetchCodeSymbols)
-      .mockImplementationOnce(() => firstCall as any)
-      .mockResolvedValueOnce({ lang: 'go', symbols: [{ name: 'Real', kind: 'function', line: 1, endLine: 5, level: 1 }] })
-
-    const wrapper = mountTocDrawer({
-      file: { name: 'a.go', content: 'package a', path: '/a.go' },
-    })
-    await nextTick()
-
-    // Change file before first fetchCodeSymbols resolves
-    await wrapper.setProps({ file: { name: 'b.go', content: 'package b', path: '/b.go' } })
-    await nextTick()
-
-    // Now resolve the first (stale) call — should be cancelled
-    resolveFirst!({ lang: 'go', symbols: [{ name: 'Stale', kind: 'function', line: 1, endLine: 5, level: 1 }] })
-    await new Promise(r => setTimeout(r, 50))
-    await nextTick()
-
-    // Second call should render "Real" from the new file
-    const items = wrapper.findAll('.toc-item')
-    expect(items.length).toBeGreaterThanOrEqual(1)
-  })
-})
-
-describe('TocDrawer — search filtering', () => {
-  it('filters toc items by search query', async () => {
-    const wrapper = mountTocDrawer({
-      file: { name: 'readme.md', content: '# Alpha\n## Beta\n## Gamma', path: '/readme.md' },
-    })
-    await nextTick()
-    await nextTick()
-
-    // Set search query
-    const input = wrapper.find('.search-input-stub')
-    await input.setValue('beta')
-    await nextTick()
-
-    const items = wrapper.findAll('.toc-item')
-    // Should only show items matching "beta"
-    expect(items.length).toBeLessThanOrEqual(3)
-  })
-})
-
-describe('TocDrawer — onBeforeUnmount', () => {
-  it('disconnects observer on unmount', async () => {
-    const disconnectSpy = vi.spyOn(MockIntersectionObserver.prototype, 'disconnect')
-    const wrapper = mountTocDrawer({
-      file: { name: 'readme.md', content: '# Test', path: '/test.md' },
-      open: true,
-    })
-    await nextTick()
-    await nextTick()
-
-    wrapper.unmount()
-    // onBeforeUnmount should disconnect observer (may or may not have been set up)
-    // At minimum, unmount should not throw
-    expect(true).toBe(true)
-    disconnectSpy.mockRestore()
+  it('forwards jumpPage from TocPanel and emits close (bottom-sheet: tap to dismiss)', () => {
+    const wrapper = mountDrawer({ file: { name: 'doc.pdf', path: '/doc.pdf' } })
+    const panel = wrapper.findComponent({ name: 'TocPanel' })
+    panel.vm.$emit('jumpPage', 3)
+    expect(wrapper.emitted('jumpPage')).toEqual([[3]])
+    expect(wrapper.emitted('close')).toBeTruthy()
   })
 })
