+++ web/src/components/file/__tests__/TocDock.test.ts
@@ -21,12 +21,12 @@ vi.mock('@/composables/useTocDockPreference', () => ({
   }),
 }))
 
-vi.mock('@/components/TocDrawer.vue', () => ({
+vi.mock('@/components/TocPanel.vue', () => ({
   default: defineComponent({
-    name: 'TocDrawer',
-    props: { open: Boolean, file: Object, pdfOutline: Array, docked: Boolean },
-    emits: ['close', 'jump', 'jumpPage'],
-    template: '<div class="toc-drawer-stub"><slot /><button class="toc-close-stub" @click="$emit(\'close\')">close</button></div>',
+    name: 'TocPanel',
+    props: { file: Object, pdfOutline: Array },
+    emits: ['jump', 'jumpPage'],
+    template: '<div class="toc-panel-stub"><button class="toc-jump-stub" @click="$emit(\'jump\', 12)">jump</button></div>',
   }),
 }))
 
@@ -52,26 +52,44 @@ describe('TocDock', () => {
     tocDockWidth.value = 260
   })
 
-  it('renders the dock container with the drawer when open', () => {
+  it('renders the dock container with the inline TocPanel when open', () => {
     const wrapper = mountDock()
     expect(wrapper.find('.toc-dock').exists()).toBe(true)
-    expect(wrapper.find('.toc-drawer-stub').exists()).toBe(true)
+    expect(wrapper.find('.toc-panel-stub').exists()).toBe(true)
   })
 
-  it('passes docked=true and the file to TocDrawer', () => {
+  it('does NOT render the BottomSheet-backed TocDrawer (inline + popup bug)', () => {
     const wrapper = mountDock()
-    const drawer = wrapper.findComponent({ name: 'TocDrawer' })
-    expect(drawer.props('docked')).toBe(true)
-    expect(drawer.props('file')).toMatchObject({ name: 'readme.md' })
+    // Regression: the dock must use the pure-content TocPanel, never the
+    // BottomSheet-backed TocDrawer — otherwise the bottom-sheet popup would
+    // appear alongside the inline dock.
+    expect(wrapper.findComponent({ name: 'TocDrawer' }).exists()).toBe(false)
+    expect(wrapper.find('.bottom-sheet').exists()).toBe(false)
+    expect(wrapper.find('.toc-panel-stub').exists()).toBe(true)
   })
 
-  it('forwards close/jump/jumpPage events from the drawer', () => {
+  it('passes the file and pdfOutline to TocPanel', () => {
+    const wrapper = mountDock({ pdfOutline: [{ id: 'p1', text: 'Page 1', level: 1, line: 1 }] })
+    const panel = wrapper.findComponent({ name: 'TocPanel' })
+    expect(panel.props('file')).toMatchObject({ name: 'readme.md' })
+    expect(panel.props('pdfOutline')).toEqual([{ id: 'p1', text: 'Page 1', level: 1, line: 1 }])
+  })
+
+  it('forwards jump/jumpPage events from TocPanel without closing (docked stays open)', () => {
     const wrapper = mountDock()
-    const drawer = wrapper.findComponent({ name: 'TocDrawer' })
-    drawer.vm.$emit('jump', 12)
-    drawer.vm.$emit('jumpPage', 3)
+    const panel = wrapper.findComponent({ name: 'TocPanel' })
+    panel.vm.$emit('jump', 12)
+    panel.vm.$emit('jumpPage', 3)
     expect(wrapper.emitted('jump')).toEqual([[12]])
     expect(wrapper.emitted('jumpPage')).toEqual([[3]])
+    // Docked mode keeps the panel open after jumping to an item.
+    expect(wrapper.emitted('close')).toBeFalsy()
+  })
+
+  it('emits close when the dock close button is clicked', async () => {
+    const wrapper = mountDock()
+    await wrapper.find('.toc-dock-close').trigger('click')
+    expect(wrapper.emitted('close')).toBeTruthy()
   })
 
   it('applies the dock width from preference', () => {
@@ -91,47 +109,44 @@ describe('TocDock', () => {
     expect(getComputedStyle(dock.element).minWidth).toBe('200px')
   })
 
-  it('sets width incrementally when the divider is dragged', async () => {
+  it('drags RIGHT to narrow the dock (divider follows pointer on left edge)', async () => {
     const wrapper = mountDock()
     const divider = wrapper.find('.toc-dock-divider')
+    pressDivider(divider, 300)
 
-    // Press at clientX=300, drag right to 340 (+40): width grows 260 → 300.
-    const down = new MouseEvent('pointerdown', { clientX: 300, bubbles: true })
-    divider.element.dispatchEvent(down)
+    // Press at clientX=300, drag right to 340 (+40): the dock's left edge
+    // follows the pointer rightward, narrowing the dock 260 → 220.
+    window.dispatchEvent(new PointerEvent('pointermove', { pointerId: 1, bubbles: true, clientX: 340 }))
     await nextTick()
 
-    window.dispatchEvent(new MouseEvent('pointermove', { clientX: 340, bubbles: true }))
-    await nextTick()
-
-    window.dispatchEvent(new MouseEvent('pointerup', { bubbles: true }))
+    window.dispatchEvent(new PointerEvent('pointerup', { pointerId: 1, bubbles: true }))
     await nextTick()
 
     expect(setWidth).toHaveBeenCalled()
     const calledWidth = setWidth.mock.calls[0][0]
-    expect(calledWidth).toBe(300)
+    expect(calledWidth).toBe(220)
   })
 
-  it('drags left to shrink the dock', async () => {
+  it('drags LEFT to widen the dock', async () => {
     const wrapper = mountDock()
     const divider = wrapper.find('.toc-dock-divider')
+    pressDivider(divider, 400)
 
-    // Press at clientX=400, drag left to 350 (−50): width shrinks 260 → 210.
-    divider.element.dispatchEvent(new MouseEvent('pointerdown', { clientX: 400, bubbles: true }))
+    // Press at clientX=400, drag left to 350 (−50): the dock's left edge moves
+    // left, widening the dock 260 → 310.
+    window.dispatchEvent(new PointerEvent('pointermove', { pointerId: 1, bubbles: true, clientX: 350 }))
     await nextTick()
 
-    window.dispatchEvent(new MouseEvent('pointermove', { clientX: 350, bubbles: true }))
-    await nextTick()
-
-    window.dispatchEvent(new MouseEvent('pointerup', { bubbles: true }))
+    window.dispatchEvent(new PointerEvent('pointerup', { pointerId: 1, bubbles: true }))
     await nextTick()
 
     expect(setWidth).toHaveBeenCalled()
-    expect(setWidth.mock.calls[0][0]).toBe(210)
+    expect(setWidth.mock.calls[0][0]).toBe(310)
   })
 
   it('ignores pointermove while not dragging', async () => {
     const wrapper = mountDock()
-    window.dispatchEvent(new MouseEvent('pointermove', { clientX: 340, bubbles: true }))
+    window.dispatchEvent(new PointerEvent('pointermove', { pointerId: 1, bubbles: true, clientX: 340 }))
     await nextTick()
     expect(setWidth).not.toHaveBeenCalled()
   })
@@ -139,7 +154,7 @@ describe('TocDock', () => {
   it('cleans up drag listeners on unmount while dragging', async () => {
     const wrapper = mountDock()
     const divider = wrapper.find('.toc-dock-divider')
-    divider.element.dispatchEvent(new MouseEvent('pointerdown', { clientX: 300, bubbles: true }))
+    pressDivider(divider, 300)
     await nextTick()
     expect(document.body.classList.contains('toc-dock-resizing')).toBe(true)
 
@@ -148,3 +163,11 @@ describe('TocDock', () => {
     expect(document.body.classList.contains('toc-dock-resizing')).toBe(false)
   })
 })
+
+/** Dispatch pointerdown on the divider with pointer-capture mocked (jsdom lacks it). */
+function pressDivider(divider: { element: Element | null }, clientX: number) {
+  const el = divider.element as HTMLElement
+  el.setPointerCapture = vi.fn()
+  el.releasePointerCapture = vi.fn()
+  el.dispatchEvent(new PointerEvent('pointerdown', { pointerId: 1, button: 0, bubbles: true, clientX }))
+}
